requirements.txt
docx==0.2.4
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
PyPDF2==3.0.1
python_docx==1.1.2
streamlit==1.35.0
wordcloud==1.9.3
